[PB-12] JsonLog.hs transferred from block to infra/Pos/Util/JsonLog/Events.hs
