+++ auxx/src/Command/BlockGen.hs
@@ -17,8 +17,9 @@ import           Pos.Core (gdBootStakeholders, genesisData)
 import           Pos.Crypto (encToSecret)
 import           Pos.Generator.Block (BlockGenParams (..), genBlocks, tgpTxCountRange)
 import           Pos.StateLock (Priority (..), withStateLock)
-import           Pos.Txp (MemPoolModifyReason (..), txpGlobalSettings)
+import           Pos.Txp (txpGlobalSettings)
 import           Pos.Util.CompileInfo (withCompileInfo)
+import           Pos.Util.JsonLog.Events (MemPoolModifyReason (..))
 
 import           Lang.Value (GenBlocksParams (..))
 import           Mode (MonadAuxxMode)
