+++ auxx/src/Command/Rollback.hs
@@ -23,9 +23,10 @@ import qualified Pos.DB.Block.Load as DB
 import qualified Pos.DB.BlockIndex as DB
 import           Pos.Ssc.Configuration (HasSscConfiguration)
 import           Pos.StateLock (Priority (..), withStateLock)
-import           Pos.Txp (MemPoolModifyReason (..), flattenTxPayload)
+import           Pos.Txp (flattenTxPayload)
 import           Pos.Util.Chrono (NewestFirst, _NewestFirst)
 import           Pos.Util.CompileInfo (HasCompileInfo)
+import           Pos.Util.JsonLog.Events (MemPoolModifyReason (..))
 
 import           Mode (MonadAuxxMode)
