+++ block/src/Pos/Block/Network/Logic.hs
@@ -8,6 +8,9 @@ module Pos.Block.Network.Logic
        (
          BlockNetLogicException (..)
        , triggerRecovery
+       , requestTipOuts
+       , requestTip
+
        , handleBlocks
 
        , handleUnsolicitedHeader
@@ -17,27 +20,27 @@ import           Universum
 
 import           Control.Concurrent.STM (isFullTBQueue, readTVar, writeTBQueue, writeTVar)
 import           Control.Exception.Safe (Exception (..))
-import           Control.Exception (IOException)
 import qualified Data.List.NonEmpty as NE
-import qualified Data.Map.Strict as M
 import qualified Data.Text.Buildable as B
 import           Formatting (bprint, build, sformat, shown, stext, (%))
-import           Mockable (forConcurrently)
 import           Serokell.Util.Text (listJson)
 import qualified System.Metrics.Gauge as Metrics
 import           System.Wlog (logDebug, logInfo, logWarning)
 
 import           Pos.Binary.Txp ()
-import           Pos.Block.BlockWorkMode (BlockWorkMode)
+import           Pos.Block.BlockWorkMode (BlockInstancesConstraint, BlockWorkMode)
 import           Pos.Block.Configuration (criticalForkThreshold)
 import           Pos.Block.Error (ApplyBlocksException)
 import           Pos.Block.Logic (ClassifyHeaderRes (..), classifyNewHeader, lcaWithMainChain,
                                   verifyAndApplyBlocks)
 import qualified Pos.Block.Logic as L
+import           Pos.Block.Network.Types (MsgGetHeaders (..), MsgHeaders (..))
 import           Pos.Block.RetrievalQueue (BlockRetrievalQueue, BlockRetrievalQueueTag,
                                            BlockRetrievalTask (..))
 import           Pos.Block.Types (Blund, LastKnownHeaderTag)
-import           Pos.Communication.Protocol (NodeId)
+import           Pos.Communication.Limits.Types (recvLimited)
+import           Pos.Communication.Protocol (ConversationActions (..), NodeId, OutSpecs, convH,
+                                             toOutSpecs)
 import           Pos.Core (HasHeaderHash (..), HeaderHash, gbHeader, headerHashG, isMoreDifficult,
                            prevBlockL)
 import           Pos.Core.Block (Block, BlockHeader, blockHeader)
@@ -49,13 +52,12 @@ import           Pos.Exception (cardanoExceptionFromException, cardanoExceptionT
 import           Pos.Recovery.Info (recoveryInProgress)
 import           Pos.Reporting.MemState (HasMisbehaviorMetrics (..), MisbehaviorMetrics (..))
 import           Pos.Reporting.Methods (reportMisbehaviour)
-import           Pos.StateLock (Priority (..), modifyStateLock, withStateLockNoMetrics)
-import           Pos.Txp.MemState.Types (MemPoolModifyReason (..))
-import           Pos.Util (buildListBounds, multilineBounds, _neHead, _neLast)
+import           Pos.StateLock (Priority (..), modifyStateLock)
+import           Pos.Util (buildListBounds, multilineBounds, _neLast)
 import           Pos.Util.AssertMode (inAssertMode)
 import           Pos.Util.Chrono (NE, NewestFirst (..), OldestFirst (..), _NewestFirst,
                                   _OldestFirst)
-import           Pos.Util.JsonLog (jlAdoptedBlock)
+import           Pos.Util.JsonLog.Events (MemPoolModifyReason (..), jlAdoptedBlock)
 import           Pos.Util.TimeWarp (CanJsonLog (..))
 import           Pos.Util.Util (lensOf)
 
@@ -98,28 +100,36 @@ triggerRecovery
     => Diffusion m -> m ()
 triggerRecovery diffusion = unlessM recoveryInProgress $ do
     logDebug "Recovery triggered, requesting tips from neighbors"
-    -- The 'catch' here is for an exception when trying to enqueue the request.
-    -- In 'requestTipsAndProcess', IO exceptions are caught, for each
-    -- individual request per-peer. Those are not re-thrown.
-    void requestTipsAndProcess `catch`
+    -- I know, it's not unsolicited. TODO rename.
+    void (Diffusion.requestTip diffusion $ handleUnsolicitedHeader) `catch`
         \(e :: SomeException) -> do
            logDebug ("Error happened in triggerRecovery: " <> show e)
            throwM e
     logDebug "Finished requesting tips for recovery"
+
+requestTipOuts :: BlockInstancesConstraint m => Proxy m -> OutSpecs
+requestTipOuts _ =
+    toOutSpecs [ convH (Proxy :: Proxy MsgGetHeaders)
+                       (Proxy :: Proxy MsgHeaders) ]
+
+-- | Is used if we're recovering after offline and want to know what's
+-- current blockchain state. Sends "what's your current tip" request
+-- to everybody we know.
+requestTip
+    :: BlockWorkMode ctx m
+    => NodeId
+    -> ConversationActions MsgGetHeaders MsgHeaders m
+    -> m ()
+requestTip nodeId conv = do
+    logDebug "Requesting tip..."
+    send conv (MsgGetHeaders [] Nothing)
+    whenJustM (recvLimited conv) handleTip
   where
-    requestTipsAndProcess = do
-        requestsMap <- Diffusion.requestTip diffusion
-        forConcurrently (M.toList requestsMap) $ \it@(nodeId, _) -> waitAndProcessOne it `catch`
-            -- Catch and squelch IOExceptions so that one failed request to one
-            -- particlar peer does not stop the others.
-            \(e :: IOException) ->
-                logDebug $ sformat ("Error requesting tip from "%shown%": "%shown) nodeId e
-    waitAndProcessOne (nodeId, mbh) = do
-        -- 'mbh' is an 'm' term that returns when the header has been
-        -- downloaded.
-        bh <- mbh
-        -- I know, it's not unsolicited. TODO rename.
-        handleUnsolicitedHeader bh nodeId
+    handleTip (MsgHeaders (NewestFirst (tip:|[]))) = do
+        logDebug $ sformat ("Got tip "%shortHashF%", processing") (headerHash tip)
+        handleUnsolicitedHeader tip nodeId
+    handleTip t =
+        logWarning $ sformat ("requestTip: got unexpected response: "%shown) t
 
 ----------------------------------------------------------------------------
 -- Headers processing
