+++ txp/Pos/Txp/Logic/Local.hs
@@ -25,7 +25,8 @@ import           Data.Default (Default (def))
 import qualified Data.HashMap.Strict as HM
 import           Formatting (build, sformat, (%))
 import           JsonLog (CanJsonLog (..))
-import           System.Wlog (NamedPureLogger, WithLogger, logDebug, logError, logWarning)
+import           System.Wlog (NamedPureLogger, WithLogger, launchNamedPureLog, logDebug, logError,
+                              logWarning)
 
 import           Pos.Core (BlockVersionData, EpochIndex, HeaderHash, siEpoch)
 import           Pos.Core.Txp (TxAux (..), TxId, TxUndo)
@@ -35,21 +36,22 @@ import qualified Pos.DB.GState.Common as GS
 import           Pos.Reporting (reportError)
 import           Pos.Slotting (MonadSlots (..))
 import           Pos.StateLock (Priority (..), StateLock, StateLockMetrics, withStateLock)
-import           Pos.Txp.MemState (GenericTxpLocalData (..), GenericTxpLocalDataPure, MempoolExt,
-                                   MemPoolModifyReason (..),
-                                   MonadTxpMem, TxpLocalWorkMode, askTxpMem, getLocalTxsMap,
-                                   getUtxoModifier, modifyTxpLocalData, setTxpLocalData)
-import           Pos.Txp.Toil (DBToil, GenericToilModifier (..), MonadUtxoRead (..), ToilT,
-                               ToilVerFailure (..), Utxo, mpLocalTxs, normalizeToil, processTx,
-                               runDBToil, runToilTLocal, runToilTLocalExtra, utxoGetReader)
+import           Pos.Txp.Logic.Common (buildUtxo)
+import           Pos.Txp.MemState (GenericTxpLocalData (..), MempoolExt, MonadTxpMem,
+                                   TxpLocalWorkMode, getLocalTxsMap, getLocalUndos, getMemPool,
+                                   getTxpExtra, getUtxoModifier, setTxpLocalData, withTxpLocalData,
+                                   withTxpLocalDataLog)
+import           Pos.Txp.Toil (ExtendedLocalToilM, LocalToilState (..), MemPool,
+                               ToilVerFailure (..), UndoMap, Utxo, UtxoLookup, UtxoModifier,
+                               extendLocalToilM, mpLocalTxs, normalizeToil, processTx, utxoToLookup)
 import           Pos.Txp.Topsort (topsortTxs)
+import           Pos.Util.JsonLog.Events (MemPoolModifyReason (..))
 import           Pos.Util.Util (HasLens')
 
 type TxpProcessTransactionMode ctx m =
     ( TxpLocalWorkMode ctx m
     , HasLens' ctx StateLock
     , HasLens' ctx (StateLockMetrics MemPoolModifyReason)
-    , MonadMask m
     , MempoolExt m ~ ()
     , CanJsonLog m
     )
