+++ wallet-new/src/Cardano/Wallet/Kernel/Mode.hs
@@ -34,7 +34,7 @@ import           Pos.Txp.Logic
 import           Pos.Txp.MemState
 import           Pos.Util
 import           Pos.Util.Chrono
-import           Pos.Util.JsonLog
+import           Pos.Util.JsonLog.Events
 import           Pos.Util.TimeWarp (CanJsonLog (..))
 import           Pos.WorkMode
