+++ wallet/cardano-sl-wallet.cabal
@@ -133,6 +133,7 @@ library
                       , data-default
                       , directory
                       , dlist
+                      , ekg-core
                       , ether
                       , exceptions
                       , filepath
